Split CRM extensions to extra
